third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/breakpoint-manager-expected.txt

Tests BreakpointManager class.


Running: testBreakpointInCollectedReload
  Created breakpoints manager
  Dumping Storage

  Adding file without script:
  Adding UISourceCode: a.js

  Setting breakpoint:
  Setting breakpoint at a.js:10:0 enabled:true condition:
    breakpointAdded(a.js, 10, 0, , true)
    debuggerModel.setBreakpoint(a.js:10:)
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)

  Reloading:
  Resetting debugger.
  Resetting workspace.
    breakpointRemoved(a.js, 10, 0)

  Adding file with script:
  Adding script: a.js
  Adding UISourceCode: a.js
    breakpointAdded(a.js, 10, 0, , true)

  Emulating breakpoint resolved event:
    breakpointRemoved(a.js, 10, 0)
    breakpointAdded(a.js, 10, 0, , true)
    Location created: a.js:10

  Make sure we don't do any unnecessary breakpoint actions:
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)
  Dumping Storage
    a.js:10 enabled:true condition:
  Resetting breakpoint manager
    breakpointRemoved(a.js, 10, 0)
    Location disposed: a.js:10
    debuggerModel.removeBreakpoint(a.js:10)
  Dumping Breakpoint Locations
